chore: Rename bytes -> bits
