+++ pkg/bits/bits.go
@@ -1,4 +1,4 @@
-package bytes
+package bits
 
 func SplitU16(value uint16) (uint8, uint8) {
 	return uint8(value >> 8), uint8(value & 0b1111_1111)
